\name{cacheR-package}
\alias{cacheR-package}
\alias{cacheR}
\docType{package}
\title{
\packageTitle{cacheR}
}
\description{
\packageDescription{cacheR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cacheR}
\packageIndices{cacheR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cacheR}

Maintainer: \packageMaintainer{cacheR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
